+++ calcuMLator/data.py
@@ -18,12 +18,7 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 '''
 import numpy as np
-
-# data sets size
-TRAINING_SIZE = 100
-TRAINING_STEP = 3
-TEST_SIZE = 1600
-TEST_RANGE = 10**3
+from sklearn.utils import shuffle
 
 def create_random_set(size, maximum):
     '''
@@ -45,7 +40,8 @@ def create_full_set(step, maximum):
     Returns a grid set with up to the "maximum" value and with "step" spacing
     '''
     X, y_add, y_sub, y_mul, y_div = [], [], [], [], []
-    dat1 = np.logspace(-1, step, )
+    dat1 = np.logspace(-1, step, maximum)
+    # dat1 = np.linspace(0, 5, 10)
     x0 = np.append(-np.flipud(dat1), dat1)
     for i in x0:
         for j in x0:
@@ -56,11 +52,5 @@ def create_full_set(step, maximum):
             y_sub.append(np.subtract(i, j))
             y_mul.append(np.multiply(i, j))
             y_div.append(np.divide(i, j))
-    return np.array(X), np.array(y_add), np.array(y_sub), np.array(y_mul), np.array(y_div)
 
-# create training set
-X_train, y_train_add, y_train_sub, y_train_mul, y_train_div= create_full_set(TRAINING_STEP, TRAINING_SIZE)
-
-# create test set
-X_test, y_test_add, y_test_sub, y_test_mul, y_test_div = create_random_set(TEST_RANGE, TEST_SIZE)
-print(X_train.shape)
+    return shuffle(np.array(X), np.array(y_add), np.array(y_sub), np.array(y_mul), np.array(y_div))