+++ calcuMLator/train.py
@@ -22,14 +22,30 @@ from collections import OrderedDict
 from os import path
 import numpy as np
 from sklearn import svm, grid_search, metrics, linear_model, ensemble
+from sklearn.gaussian_process import GaussianProcess
+from sklearn.neural_network import BernoulliRBM
 from sklearn.externals import joblib
-from data import *
+import data
 
 FOLDER = path.dirname(path.realpath(__file__))+'/'
 ESTIMATOR_FOLDER = FOLDER+'estimators/'
 ESTIMATOR_CONF = FOLDER+'estimator_conf.json'
 with open(ESTIMATOR_CONF) as text:
     conf = json.loads(text.read())
+
+# data sets size
+TRAINING_SIZE = 22
+TRAINING_STEP = 3
+TEST_SIZE = 10**3
+TEST_RANGE = 10**3
+
+# create training set
+training_set = data.create_full_set(TRAINING_STEP, TRAINING_SIZE)
+X_train, y_train_add, y_train_sub, y_train_mul, y_train_div = training_set
+
+# create test set
+test_set = data.create_random_set(TEST_RANGE, TEST_SIZE)
+X_test, y_test_add, y_test_sub, y_test_mul, y_test_div = test_set
 
 
 def train_linear(X_train, y_train, X_test, y_test):
@@ -38,26 +54,9 @@ def train_linear(X_train, y_train, X_test, y_test):
     '''
     clf_linear = linear_model.LinearRegression()
     clf_linear.fit(X_train, y_train)
-    r2_linear = metrics.r2_score(y_test, clf_linear.predict(X_test))
-    return clf_linear, r2_linear
-
-def train_SVR(X_train, y_train, X_test, y_test):
-    '''
-    Creates the rbf estimator for SVR and returns it along with it's r2_score
-    '''
-    clf_SVR = svm.SVR()
-    clf_SVR.fit(X_train, y_train)
-    r2_SVR = metrics.r2_score(y_test, clf_SVR.predict(X_test))
-    return clf_SVR, r2_SVR
-
-def train_bagging(X_train, y_train, X_test, y_test):
-    '''
-    Creates ensemble regressor estimator and returns it along with it's r2_score
-    '''
-    clf_bagging = ensemble.BaggingRegressor()
-    clf_bagging.fit(X_train, y_train)
-    r2_bagging = metrics.r2_score(y_test, clf_bagging.predict(X_test))
-    return clf_bagging, r2_bagging
+    r2_linear = (metrics.r2_score(y_test, clf_linear.predict(X_test)), metrics.r2_score(y_train, clf_linear.predict(X_train)))
+    coef_linear = {'coef': clf_linear.coef_.tolist(), 'intercept': clf_linear.intercept_}
+    return clf_linear, r2_linear, coef_linear
 
 def train_ridge(X_train, y_train, X_test, y_test):
     '''
@@ -65,31 +64,44 @@ def train_ridge(X_train, y_train, X_test, y_test):
     '''
     clf_ridge = linear_model.RidgeCV()
     clf_ridge.fit(X_train, y_train)
-    r2_ridge = metrics.r2_score(y_test, clf_ridge.predict(X_test))
-    return clf_ridge, r2_ridge
+    r2_ridge = metrics.r2_score(y_test, clf_ridge.predict(X_test)), metrics.r2_score(y_train, clf_ridge.predict(X_train))
+    coef_ridge = {'coef': clf_ridge.coef_.tolist(), 'intercept': clf_ridge.intercept_, 'aplpha': clf_ridge.alpha_}
+    return clf_ridge, r2_ridge, coef_ridge
 
-def train_bayesian(X_train, y_train, X_test, y_test):
+def train_SVR(X_train, y_train, X_test, y_test):
     '''
-    Creates ridge regression estimator and returns it along with it's r2_score
+    Creates the rbf estimator for SVR and returns it along with it's r2_score
     '''
-    clf_bayesian = linear_model.BayesianRidge()
-    clf_bayesian.fit(X_train, y_train)
-    r2_bayesian = metrics.r2_score(y_test, clf_bayesian.predict(X_test))
-    return clf_bayesian, r2_bayesian
+    # xtrain, ytrain = X_train[:200], y_train[:200]
+    clf_SVR = svm.SVR(C= 3**10, epsilon= 3**-2, kernel='rbf')
+    clf_SVR.fit(X_train, y_train)
+    r2_SVR = metrics.r2_score(y_test, clf_SVR.predict(X_test)), metrics.r2_score(y_train, clf_SVR.predict(X_train))
+    coef_SVR = {'support_vectors': len(clf_SVR.support_vectors_), 'intercept': clf_SVR.intercept_.tolist()}
+    return clf_SVR, r2_SVR, coef_SVR
 
-def train_lars(X_train, y_train, X_test, y_test):
+def train_bagging(X_train, y_train, X_test, y_test):
     '''
-    Creates ridge regression estimator and returns it along with it's r2_score
+    Creates bagging regressor estimator and returns it along with it's r2_score
     '''
-    clf_lars = linear_model.LarsCV()
-    clf_lars.fit(X_train, y_train)
-    r2_lars = metrics.r2_score(y_test, clf_lars.predict(X_test))
-    return clf_lars, r2_lars
+    clf_bagging = ensemble.BaggingRegressor()
+    clf_bagging.fit(X_train, y_train)
+    r2_bagging = metrics.r2_score(y_test, clf_bagging.predict(X_test)), metrics.r2_score(y_train, clf_bagging.predict(X_train))
+    coef_bagging = {'estimators': len(clf_bagging.estimators_), 'estimators_features': len(clf_bagging.estimators_features_)}
+    return clf_bagging, r2_bagging, coef_bagging
 
+def train_gaussian(X_train, y_train, X_test, y_test):
+    '''
+    Creates bagging regressor estimator and returns it along with it's r2_score
+    '''
+    clf_gaussian = GaussianProcess(theta0=1e-2, thetaL=1e-4, thetaU=1e-0)
+    clf_gaussian.fit(X_train[:500], y_train[:500])
+    r2_gaussian = metrics.r2_score(y_test, clf_gaussian.predict(X_test)), metrics.r2_score(y_train, clf_gaussian.predict(X_train))
+    coef_gaussian = {'theta': clf_gaussian.theta_.tolist()}
+    return clf_gaussian, r2_gaussian, coef_gaussian
 
 def train_MLP(X_train, y_train, X_test, y_test):
     '''
-    Creates the MPL estimator and returns it along with it's r2_score
+    Creates the MLP estimator and returns it along with it's r2_score
     '''
 
 def train_all():
@@ -97,22 +109,29 @@ def train_all():
     Trains all the estimators
     '''
     report = OrderedDict()
+    coef = OrderedDict()
+    # conf['estimators'] = ['linear', 'ridge', 'gaussian', 'bagging', 'SVR']
+    conf['estimators'] = []
     for estimator in conf['estimators']:
         report[estimator] = OrderedDict()
+        coef[estimator] = OrderedDict()
         for operator in conf['types'].values():
             # arguments - inject your code here!
             args = (X_train, eval('y_train_'+operator), X_test, eval('y_test_'+operator))
             estimator_path = ESTIMATOR_FOLDER+operator+'_'+conf[estimator]
 
             # create estimator - two evals?? seriously??
-            clf, r2 = eval('train_'+estimator+'(*args)')
-            print(estimator+' '+operator+' trained! r2 score: '+str(r2))
+            clf, r2, c= eval('train_'+estimator+'(*args)')
+            print(estimator+' '+operator+' trained! r2 score (test, train): '+str(r2))
 
             # save estimator
             joblib.dump(clf, estimator_path)
             report[estimator][operator] = r2
+            coef[estimator][operator] = c
     with open(FOLDER+conf['report'], 'w') as text:
         text.write(json.dumps(report, indent=2))
+    with open(FOLDER+conf['coef'], 'w') as text:
+        text.write(json.dumps(coef, indent=2))
 
 if __name__ == '__main__':
     train_all()